+++ Transports/com.community.netcode.transport.webrtc/Signaling~/src/server.config.ts
@@ -1,6 +1,0 @@
-export const serverConfig = {
-  port: +(process.env.PORT || 4000),
-  auth: {
-    token: process.env.AUTH_TOKEN || '',
-  },
-};